Fix return type of sys_lseek()
-Syscall sys_lseek and other device specific lseek functions
 returned signed 32 bit type loff_t. Changed to return int.

The kernel was compiled with BCC and tested with QEMU. Works.
Code size reduced by 64 bytes.
IA16-GCC produced kernel works with QEMU running under 32 bit
Debian 6.0 and 64 bit Ubuntu 13.10.

+++ elks/fs/pipe.c
@@ -45,7 +45,7 @@ int get_unused_fd(struct file *f)
     return -EMFILE;
 }
 
-loff_t pipe_lseek(struct inode *inode, struct file *file, loff_t offset,
+int pipe_lseek(struct inode *inode, struct file *file, loff_t offset,
 		  int orig)
 {
     debug("PIPE: lseek called.\n");
